refactor: cascade cleanup wave 3 — NativeScripts, Governance, Metadata, Plutus, Block, Interval
15 more files brought green via 3 parallel cleanup passes:

- NativeScript/NativeScript.swift, ScriptNofK.swift, AfterScript.swift:
  AfterScript.slot widened from Int → SlotNumber (consistency with
  BeforeScript landed earlier). NativeScript TYPE rawValue cascades
  cleaned.
- Governance/Voting.swift, GovernanceActions/UpdateCommittee.swift,
  TreasuryWithdrawalsAction.swift: discriminator code casts at boundary.
- Transaction/Datum.swift: DatumOption + RawDatum + Datum boundary
  casts; type field stays Int (Plutus version discriminator).
- StakePool/OperationalCertificateIssueCounter.swift, SingleHostName.swift:
  redundant UInt() drops + UInt64() at boundary.
- Certificates/VoteRegisterDelegate.swift: same pattern.
- Metadata/Metadata.swift: bounds-checked Int64 → Int narrowing on the
  .int branch (matches the pre-existing .uint branch overflow guard).
- Types/Interval.swift: NonNegativeInterval.lowerBound widened
  UInt → UInt64 (struct field type change — semantically already held
  UInt64-range values via doubleToNonNegativeInterval). Also touches
  ProtocolParameters.swift L24 to match the new constructor signature.
- Block/Block.swift, Plutus/PlutusTypes.swift, Plutus/PlutusData/Constr.swift:
  Plutus version field stays Int (small discriminator); other sites
  cast at boundary.

Total errors 4052 → 2112 (~1940 cascade resolved). 43 files still red.

+++ Sources/SwiftCardanoCore/Block/Block.swift
@@ -161,11 +161,11 @@ public struct Block: Serializable {
         // auxiliary_data_set
         var auxDataDict = OrderedDictionary<Primitive, Primitive>()
         for (index, auxData) in auxiliaryDataSet {
-            auxDataDict[.uint(UInt(index))] = try auxData.toPrimitive()
+            auxDataDict[.uint(UInt64(index))] = try auxData.toPrimitive()
         }
 
         // invalid_transactions
-        let invalidTxs: [Primitive] = invalidTransactions.map { .uint(UInt($0)) }
+        let invalidTxs: [Primitive] = invalidTransactions.map { .uint(UInt64($0)) }
 
         return .list([
             try header.toPrimitive(),
@@ -274,12 +274,12 @@ public struct Block: Serializable {
 
         var auxDict = OrderedDictionary<Primitive, Primitive>()
         for (index, auxData) in auxiliaryDataSet {
-            auxDict[.uint(UInt(index))] = try auxData.toDict()
+            auxDict[.uint(UInt64(index))] = try auxData.toDict()
         }
         dict[.string(CodingKeys.auxiliaryDataSet.rawValue)] = .orderedDict(auxDict)
 
         dict[.string(CodingKeys.invalidTransactions.rawValue)] = .list(
-            invalidTransactions.map { .uint(UInt($0)) }
+            invalidTransactions.map { .uint(UInt64($0)) }
         )
 
         return .orderedDict(dict)
